handle DATABASE_NAME or DATABASE_URL

+++ restore.sh
@@ -16,6 +16,10 @@ case $key in
     DATABASE_NAME="$2"
     shift
     ;;
+    -url|--dburl)
+    DATABASE_URL="$2"
+    shift
+    ;;
 esac
 shift
 done
@@ -23,6 +27,7 @@ done
 BUCKET_NAME=${BUCKET_NAME:='bucket_name'}
 BUCKET_PATH=${BUCKET_PATH:='bucket_path'}
 DATABASE_NAME=${DATABASE_NAME:='database_name'}
+DATABASE_URL=${DATABASE_URL:=}
 DB_BACKUP_ENC_KEY=${DB_BACKUP_ENC_KEY:=}
 
 if [[ -z "$DATABASE_NAME" ]]; then
@@ -87,34 +92,59 @@ openssl enc -aes-256-cbc -d -pass "env:DB_BACKUP_ENC_KEY" \
   -in ./$LATEST_TAR_FORMAT \
   -out ./${DATABASE_NAME}_tar_format.tar.gz
 
-# custom format
-createdb ${DATABASE_NAME}_restored
-time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_custom_format.dump
-psql --dbname ${DATABASE_NAME}_restored --command "\d"
-dropdb ${DATABASE_NAME}_restored
+if [[ -n "$DATABASE_URL" ]]; then
+  echo "Restore database using DATABASE_URL"
+  # custom format
+  time pg_restore --no-owner --dbname $DATABASE_URL ./${DATABASE_NAME}_custom_format.dump
+  psql --dbname $DATABASE_URL --command "\d"
 
-# directory format
-tar -zxvf ./${DATABASE_NAME}_directory_format.tar.gz -s "/${DATABASE_NAME}.*_directory_format/${DATABASE_NAME}_directory_format/"
-createdb ${DATABASE_NAME}_restored
-time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_directory_format
-psql --dbname ${DATABASE_NAME}_restored --command "\d"
-dropdb ${DATABASE_NAME}_restored
+  # directory format
+  tar -zxvf ./${DATABASE_NAME}_directory_format.tar.gz -s "/${DATABASE_NAME}.*_directory_format/${DATABASE_NAME}_directory_format/"
+  time pg_restore --no-owner --dbname $DATABASE_URL ./${DATABASE_NAME}_directory_format
+  psql --dbname $DATABASE_URL --command "\d"
 
-# plain format
-gzip --verbose --decompress ./${DATABASE_NAME}_plain_format.sql.gz
-UNZIPPED_FILENAME=${DATABASE_NAME}_plain_format.sql
-createdb ${DATABASE_NAME}_restored
-time psql --dbname ${DATABASE_NAME}_restored --file=$UNZIPPED_FILENAME
-psql --dbname ${DATABASE_NAME}_restored --command "\d"
-dropdb ${DATABASE_NAME}_restored
+  # plain format
+  gzip --verbose --decompress ./${DATABASE_NAME}_plain_format.sql.gz
+  UNZIPPED_FILENAME=${DATABASE_NAME}_plain_format.sql
+  time psql --dbname $DATABASE_URL --file=$UNZIPPED_FILENAME
+  psql --dbname $DATABASE_URL --command "\d"
 
-# tar format
-gzip --verbose --decompress ./${DATABASE_NAME}_tar_format.tar.gz
-UNZIPPED_FILENAME=${DATABASE_NAME}_tar_format.tar
-createdb ${DATABASE_NAME}_restored
-time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_tar_format.tar
-psql --dbname ${DATABASE_NAME}_restored --command "\d"
-dropdb ${DATABASE_NAME}_restored
+  # tar format
+  gzip --verbose --decompress ./${DATABASE_NAME}_tar_format.tar.gz
+  UNZIPPED_FILENAME=${DATABASE_NAME}_tar_format.tar
+  time pg_restore --no-owner --dbname $DATABASE_URL ./${DATABASE_NAME}_tar_format.tar
+  psql --dbname $DATABASE_URL --command "\d"
+elif [[ -n "$DATABASE_NAME" ]]; then
+  echo "Restore database using DATABASE_NAME"
+  # custom format
+  createdb ${DATABASE_NAME}_restored
+  time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_custom_format.dump
+  psql --dbname ${DATABASE_NAME}_restored --command "\d"
+  dropdb ${DATABASE_NAME}_restored
+
+  # directory format
+  tar -zxvf ./${DATABASE_NAME}_directory_format.tar.gz -s "/${DATABASE_NAME}.*_directory_format/${DATABASE_NAME}_directory_format/"
+  createdb ${DATABASE_NAME}_restored
+  time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_directory_format
+  psql --dbname ${DATABASE_NAME}_restored --command "\d"
+  dropdb ${DATABASE_NAME}_restored
+
+  # plain format
+  gzip --verbose --decompress ./${DATABASE_NAME}_plain_format.sql.gz
+  UNZIPPED_FILENAME=${DATABASE_NAME}_plain_format.sql
+  createdb ${DATABASE_NAME}_restored
+  time psql --dbname ${DATABASE_NAME}_restored --file=$UNZIPPED_FILENAME
+  psql --dbname ${DATABASE_NAME}_restored --command "\d"
+  dropdb ${DATABASE_NAME}_restored
+
+  # tar format
+  gzip --verbose --decompress ./${DATABASE_NAME}_tar_format.tar.gz
+  UNZIPPED_FILENAME=${DATABASE_NAME}_tar_format.tar
+  createdb ${DATABASE_NAME}_restored
+  time pg_restore --no-owner --dbname ${DATABASE_NAME}_restored ./${DATABASE_NAME}_tar_format.tar
+  psql --dbname ${DATABASE_NAME}_restored --command "\d"
+  dropdb ${DATABASE_NAME}_restored
+fi
 
 # cleanup
 rm -v ${DATABASE_NAME}_*.enc
